Update test app to ensure nested option fields are properly rendered in the UI
