+++ Tests/TestCommon/Utilities/TestExtensions.swift
@@ -11,6 +11,8 @@
  */
 
 import Foundation
+import XCTest
+@testable import OktaIdx
 
 extension Data {
     func urlFormEncoded() -> [String:String?]? {
@@ -30,5 +32,20 @@ extension String {
     func isBase64URLEncoded() -> Bool {
         let charset = CharacterSet(charactersIn: "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_").inverted
         return (rangeOfCharacter(from: charset) == nil)
+    }
+}
+
+extension XCTestCase {
+    func data(for json: String) -> Data {
+        return json.data(using: .utf8)!
+    }
+    
+    func decode<T>(type: T.Type, _ json: String) throws -> T where T : Decodable {
+        let jsonData = data(for: json)
+        return try JSONDecoder.idxResponseDecoder.decode(T.self, from: jsonData)
+    }
+
+    func decode<T>(type: T.Type, _ json: String, _ test: ((T) throws -> Void)) throws where T : Decodable {
+        try test(try decode(type: type, json))
     }
 }
